content_scraper.py
import asyncio
import logging
import feedparser
import aiohttp
from datetime import datetime
import re
from typing import List, Dict
from bs4 import BeautifulSoup
import html
import random

logger = logging.getLogger("content_scraper")
logging.basicConfig(level=logging.INFO)

class ContentScraper:
    def __init__(self):
        pass
        
    async def scrape_content(self):
        """Placeholder for content scraping functionality"""
        return []  # Return empty list for now 