requirements.txt
# Core networking and async
aiohttp==3.9.1
asyncio==3.4.3
requests==2.31.0

# ATProto client for Bluesky
atproto==0.0.31

# LLM and AI
langchain==0.1.0
langchain-groq==0.0.1
groq==0.4.2

# RSS and content parsing
feedparser==6.0.10
beautifulsoup4==4.12.2
html5lib==1.1

# Date and time handling
python-dateutil==2.8.2
pytz==2024.1

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging and error handling
structlog==24.1.0

# HTML processing
lxml==4.9.3